unifying RootLayout
