+++ packages/next-app/src/components/RootLayout.tsx
@@ -1,0 +1,76 @@
+"use client";
+
+import { darkTheme, RainbowKitProvider, Theme } from "@rainbow-me/rainbowkit";
+import merge from "lodash.merge";
+import Image from "next/image";
+import { usePathname } from "next/navigation";
+import { WagmiConfig } from "wagmi";
+
+import ConnectWalletImage from "#/assets/connect-wallet.svg";
+import { chains, client } from "#/wagmi/client";
+
+export function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <WagmiConfig client={client}>
+      <RainbowKitProvider
+        chains={chains}
+        modalSize="compact"
+        theme={CustomTheme}
+      >
+        {children}
+      </RainbowKitProvider>
+    </WagmiConfig>
+  );
+}
+
+export function WalletNotConnectedState() {
+  const pathname = usePathname();
+  const message =
+    pathname === "/metadata"
+      ? "metadata pools"
+      : pathname === "/daoadmin"
+      ? "actions"
+      : "";
+
+  return (
+    <div className="flex h-screen w-full flex-col items-center justify-center bg-gray-900">
+      <h1 className="flex h-12 items-center text-center text-3xl font-medium not-italic text-gray-400">
+        Your {message} will appear here
+      </h1>
+      <h1 className="mb-4 flex h-12 items-center text-center	 text-3xl font-medium not-italic text-yellow-300">
+        Please, connect your wallet
+      </h1>
+      <Image src={ConnectWalletImage} height={500} width={500} alt="" />
+    </div>
+  );
+}
+
+const CustomTheme = merge(darkTheme(), {
+  colors: {
+    accentColor: "#3182CE",
+    accentColorForeground: "#E2E8F0",
+    closeButtonBackground: "#E2E8F0",
+    connectButtonText: "#E2E8F0",
+    closeButton: "#2D3748",
+    connectButtonTextError: "#E2E8F0",
+    connectionIndicator: "#30E000",
+    generalBorder: "#E2E8F0",
+    menuItemBackground: "#353d49",
+    modalBackground: "#2D3748",
+    modalBorder: "transparent",
+    modalText: "#E2E8F0",
+    modalTextDim: "#E2E8F0",
+    modalTextSecondary: "#E2E8F0",
+    selectedOptionBorder: "rgba(60, 66, 66, 0.1)",
+    standby: "#FFD641",
+  },
+  fonts: {},
+  radii: {
+    menuButton: "25px",
+    modal: "25px",
+  },
+  shadows: {
+    selectedOption: "0px 4px 8px rgba(0, 0, 0, 0.24)",
+    selectedWallet: "0px 4px 8px rgba(0, 0, 0, 0.12)",
+  },
+} as Theme);
